refactor: 보너스 번호 검증 객체에서 보너스 입력 검증을 분리
